requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap4==2.1.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
Pillow==7.2.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
